[*] причесал код, переработал структуру файлов сайдбара статистики
