+++ src/pages/Tracker/ChartsDrawer/components/ChartPanel/components/Chart/Chart.tsx
@@ -12,8 +12,11 @@ import { Bar, Pie } from 'react-chartjs-2';
 import { useChartsDrawerStore } from 'pages/Tracker/ChartsDrawer/store';
 import { observer } from 'mobx-react-lite';
 import { NoData, DisablingContainer, Wrapper } from './Chart.styles';
-import { Loader } from '../../../ui';
-import { NO_DATA_TEXT, NO_DATA_TEXT_EXTRA } from '../../../../config';
+import { Loader } from 'pages/Tracker/ChartsDrawer/components/ui';
+import {
+  NO_DATA_TEXT,
+  NO_DATA_TEXT_EXTRA,
+} from 'pages/Tracker/ChartsDrawer/config';
 
 ChartJS.register(
   ArcElement,
