+++ src/pages/Tracker/ChartsDrawer/components/ChartPanel/components/Chart/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Chart';
+export { default as Chart } from './Chart';
