+++ src/pages/Tracker/ChartsDrawer/components/ChartPanel/components/Contol/components/DatePicker/index.ts
@@ -1,1 +1,1 @@
-export { default } from './DatePicker';
+export { default as DatePicker } from './DatePicker';
