+++ src/pages/Tracker/ChartsDrawer/components/ChartPanel/components/Contol/components/RangePicker/index.ts
@@ -1,1 +1,1 @@
-export { default } from './RangePicker';
+export { default as RangePicker } from './RangePicker';
