+++ src/pages/Tracker/ChartsDrawer/components/ChartPanel/components/Contol/components/SimpleDatesSelector/index.ts
@@ -1,1 +1,1 @@
-export { default } from './SimpleDatesSelector';
+export { default as SimpleDatesSelector } from './SimpleDatesSelector';
