+++ src/pages/Tracker/ChartsDrawer/components/ChartPanel/components/index.ts
@@ -1,0 +1,3 @@
+export * from './Chart';
+
+export * from './Contol';
