+++ src/pages/Tracker/ChartsDrawer/components/List/index.ts
@@ -1,1 +1,1 @@
-export { default } from './CategoriesList';
+export { default as CategoriesList } from './CategoriesList';
