+++ src/pages/Tracker/ChartsDrawer/components/ToggleIconButton/index.ts
@@ -1,1 +1,1 @@
-export { default } from './ToggleIconButton';
+export { default as ToggleIconButton } from './ToggleIconButton';
