+++ src/pages/Tracker/ChartsDrawer/config/index.ts
@@ -1,0 +1,5 @@
+export * from './common';
+
+export * from './pieChart';
+
+export * from './barChart';
