+++ src/pages/Tracker/ChartsDrawer/mock.ts
@@ -1,5 +1,3 @@
-import { CategoryType } from 'stores/TrackerStore/types';
-
 import { ApiStatsCategoryType, BarChartApiDataType } from './types';
 
 export const mockApiPieChartData: ApiStatsCategoryType[] = [
@@ -81,53 +79,3 @@ export const mockApiBarChartData: BarChartApiDataType = {
     },
   ],
 };
-
-export const mockListCategories: (Pick<
-  CategoryType,
-  'id' | 'name' | 'color'
-> & {
-  percentString?: string;
-})[] = [
-  {
-    id: '1',
-    name: 'Frontend',
-    color: 'pink',
-    percentString: '30%',
-  },
-  {
-    id: '2',
-    name: 'Backend',
-    color: 'blue',
-    percentString: '20%',
-  },
-  {
-    id: '3',
-    name: 'Useful',
-    color: 'yellow',
-    percentString: '15%',
-  },
-  {
-    id: '4',
-    name: 'Guitar',
-    color: 'green',
-    percentString: '14%',
-  },
-  {
-    id: '5',
-    name: 'Polytech',
-    color: 'pink',
-    percentString: '10%',
-  },
-  {
-    id: '6',
-    name: 'Something',
-    color: 'skyblue',
-    percentString: '7%',
-  },
-  {
-    id: '7',
-    name: 'Something2',
-    color: 'red',
-    percentString: '4%',
-  },
-];
