+++ src/pages/Tracker/ChartsDrawer/store/ChartDrawerStore/store/BarChartContoller/index.ts
@@ -1,3 +1,1 @@
 export { default as BarChartController } from './BarChartController';
-
-export * from './types';
