+++ src/pages/Tracker/ChartsDrawer/store/ChartDrawerStore/store/PieChartController/index.ts
@@ -1,3 +1,1 @@
 export { default as PieChartController } from './PieChartController';
-
-export * from './types';
