--- src/pages/Tracker/ChartsDrawer/store/ChartDrawerStore/store/types.ts
@@ -1,11 +1,0 @@
-import { DatesSelectionTypesEnum } from '../../../types';
-
-export type DatesStringSelectionType =
-  | {
-      selectionType: DatesSelectionTypesEnum.single;
-      value: string;
-    }
-  | {
-      selectionType: DatesSelectionTypesEnum.range;
-      value: [string, string];
-    };
